Handle unset 'Command.short_help'
There appears to be a bug in Click 7.0 that has the 'short_help' value
unset, meaning we see the following exceptions:

  AttributeError: 'NoneType' object has no attribute 'split'

Resolve this with a simply falling back to 'help' if 'short_help' is
unset. This means we continue to work with all versions of Click, even
though the proper fix should be in Click itself.

Signed-off-by: Stephen Finucane <stephen@that.guru>
Closes: #21